Put the locators in the properties via the interfaces, removed the Spring functions
