--- src/test/java/demo/test/models/TestDataMails.java
@@ -1,16 +1,0 @@
-package demo.test.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-/**
- * The type Test data mails.
- */
-@Getter
-@AllArgsConstructor
-public class TestDataMails {
-    private final String senderMailLogin;
-    private final String senderMailPassword;
-    private final String recipientMailLogin;
-    private final String recipientMailPassword;
-}
