+++ src/test/java/demo/test/webentities/locators/Constants.java
@@ -1,0 +1,14 @@
+package demo.test.webentities.locators;
+
+/**
+ * The type Constants.
+ */
+public class Constants {
+    public static final String LOCATORS_PROPERTIES_DIRECTORY = "locators";
+    public static final String TUT_BY_HOME_PAGE_LOCATORS = "tut_by_home_page.properties";
+    public static final String EMAIL_ACCOUNT_PAGE_LOCATORS = "email_account_page.properties";
+    public static final String AUTHORIZE_EMAIL_PAGE_LOCATORS = "authorize_email_page.properties";
+    public static final String TUT_BY_HEADER_LOCATORS = "tut_by_header.properties";
+    public static final String MAILS_FORM_LOCATORS = "mails_form.properties";
+    public static final String ACCOUNT_FORM_LOCATORS = "account_form.properties";
+}
