+++ src/test/java/steps/ApiEmailSteps.java
@@ -3,8 +3,8 @@ package steps;
 import cucumber.api.java.After;
 import cucumber.api.java.en.And;
 import cucumber.api.java.en.Given;
-import demo.test.models.Mail;
-import demo.test.models.TestDataMails;
+import demo.test.webentities.models.Mail;
+import demo.test.webentities.models.TestDataMails;
 import demo.test.utils.FactoryInitParams;
 import webdriver.driver.Browser;
 import webdriver.utils.mail.MailUtils;
