+++ src/test/java/webdriver/BaseForm.java
@@ -3,7 +3,6 @@ package webdriver;
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.NoSuchElementException;
-import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.ExpectedCondition;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import webdriver.driver.Browser;
@@ -21,16 +20,25 @@ public abstract class BaseForm extends BaseEntity {
      */
     protected BaseForm(final By locator, final String formTitle) {
         waitForPageLoad();
-        PageFactory.initElements(Browser.getDriver(), this);
+//        PageFactory.initElements(Browser.getDriver(), this);
         checkPageMainElement(locator, formTitle);
     }
 
+    /**
+     * Wait for page load.
+     */
     private void waitForPageLoad() {
         WebDriverWait wait = new WebDriverWait(Browser.getDriver(), Browser.getTimeoutForCondition());
         wait.until((ExpectedCondition<Boolean>) driver -> ((JavascriptExecutor) driver)
             .executeScript("return document.readyState").equals("complete"));
     }
 
+    /**
+     * Check page main element.
+     *
+     * @param locator   the locator
+     * @param formTitle the form title
+     */
     private void checkPageMainElement(final By locator, final String formTitle) {
         LOGGER.info(String.format("Initialization %s element locator", formTitle));
         if (!Browser.getDriver().findElement(locator).isDisplayed()) {
