--- src/test/java/webdriver/common/ProviderData.java
@@ -1,27 +1,0 @@
-package webdriver.common;
-
-import webdriver.Logger;
-
-/**
- * The type Provider data.
- */
-public interface ProviderData {
-    Logger LOGGER = Logger.getInstance();
-
-    /**
-     * Get test data object [ ].
-     *
-     * @return the object [ ]
-     */
-    Object[] getTestData(String dataBaseLocation);
-
-    /**
-     * The enum Data base type.
-     */
-    enum DataBaseType {
-        XML,
-        CSV,
-        DB,
-        XLS
-    }
-}
