+++ src/test/java/webdriver/resources/IStageEnvironment.java
@@ -1,0 +1,12 @@
+package webdriver.resources;
+
+import org.aeonbits.owner.Config;
+
+import static webdriver.resources.Constants.PROPERTIES_STAGE;
+
+@Config.Sources(value = {"classpath:" + PROPERTIES_STAGE})
+public interface IStageEnvironment extends Config {
+
+    @DefaultValue("http://${stage}")
+    String urlLoginPage();
+}
