--- src/test/java/webdriver/resources/MailEnvironment.java
@@ -1,19 +1,0 @@
-package webdriver.resources;
-
-import org.aeonbits.owner.Config;
-
-import static webdriver.resources.Constants.PROPERTIES_MAIL;
-
-@Config.Sources({"classpath:" + PROPERTIES_MAIL})
-public interface MailEnvironment extends Config {
-
-    String serv();
-
-    int port();
-
-    @Key("smtp.auth")
-    boolean smtpAuth();
-
-    @DefaultValue("Sent")
-    String seltFolder();
-}
