+++ src/test/resources/locators/account_form.properties
@@ -1,0 +1,2 @@
+main_locator=//div[contains(@class, 'head-user') or contains(@class, 'login-title')]
+user_dropdown=//div[contains(@class, 'dropdown-content')]//*[contains(@href, \''{0}\'')]
