+++ src/test/resources/locators/authorize_email_page.properties
@@ -1,0 +1,4 @@
+main_locator=//*[@class='bookBody']
+text_box_input_login=//*[@id='form']//input[@name='login']
+text_box_input_password=//*[@id='form']//input[@name='password']
+btn_sigh_in=//*[@type='submit']
