+++ src/test/resources/locators/email_account_page.properties
@@ -1,0 +1,2 @@
+main_locator=//*[@class='mail-App-Content']
+nav_bar=//div[contains(@data-key, 'left-box')]//a[contains(@href, \''{0}\'')]
