+++ src/test/resources/locators/mails_form.properties
@@ -1,0 +1,6 @@
+main_locator=//*[contains(@class, 'mail-Layout-Main')]
+last_mail=//*[contains(@class, 'mail-Layout-Main')]//*[contains(@class, 'MessagesList')]/.
+sender=//span[contains(@class, 'sender')]
+recipient=//div[contains(@class, 'user-name')]
+subject=//div[contains(@class, 'Subject')]
+text=//div[contains(@class, 'body')]
