+++ src/test/resources/locators/tut_by_header.properties
@@ -1,0 +1,2 @@
+main_locator=//*[contains(@id, '_mainmenu')]
+top_bar=//*[contains(@id, '_mainmenu')]//ul[contains(@class, 'topbar-i')]//a[contains(@href, \''{0}\'')]
