+++ src/test/resources/locators/tut_by_home_page.properties
@@ -1,0 +1,1 @@
+main_locator=//*[@class='header-logo']
